+++ History.md
@@ -4,6 +4,7 @@ Unreleased
 * added; invoice address on previews
 * added; `invoice.OriginalInvoiceNumber`
 * added; VatLocationValid to Account
+* added; Open Amount Refunds to Invoice
 
 1.1.7 (stable) / 2014-12-19
 ==================